Do not ignore exception on main loop
This closes #77.

In the previous implementation, any exception in `runLoop` was ignored
until `Gtk.main` ends. This resulted to weird behaviors where user
`update` / `view` functions where ignored, but the application was still
running and primitive gtk events (such as style change on mouse
overlays) were still running, resulting in an apparently alive but
unresponsive application. The exception was also not printed.

With this change, the application now quit if any exception is raised in
the `view` and `update` user function. User are in charge of setting an
exception handler if they want a different behavior.