Force exception to be rethrown in main thread
Exceptions from view and inputs producers may ends up ignored or throw
by the wrong thread.

This leads to the following unsatisfying behaviors:

- Ignored exceptions (from view or producers) were able to kill
important worker threads without fully killing the application. That one
was stuck in a broken state where events were not handled.
- Exception in the gtk UI thread would kill the application but by
raising an exception which is not catchable by any process, so we were
not able to unit test it OR even recover from it.

The changes are the following:

- All `forkIO` are replaced by `async` calls, mostly `concurrently` and
`mapConcurrently` which ensures that any exception in one forked thread
will kill the other and be re-raised in the main thread.
- Any exception in the GTK UI thread is reraised in the main thread.

The observable result of this is that any exception will fully kill the
application if not caught in the user view / update / inputs functions
directly OR outside of the `runLoop` function.